moved errors to their respective classes
